ci(tool-call): Move integration test to IT class
